+++ src/frontend/apps/impress/src/features/docs/doc-editor/components/index.ts
@@ -1,4 +1,3 @@
 export * from './BlockNoteEditor';
 export * from './DocEditor';
-export * from './EmojiPicker';
 export * from './custom-blocks/';